requirements.txt
requests==2.19.1
selenium==2.48.0
pandas==0.23.4
pytesseract==0.2.5
Pillow==5.3.0
config==0.4.0
